Refactored score calculator
