+++ src/main/java/com/idealista/domain/conditions/Rule.java
@@ -1,0 +1,8 @@
+package com.idealista.domain.conditions;
+
+import com.idealista.domain.Ad;
+
+public interface Rule {
+
+    Ad apply(Ad ad);
+}
